Split vmbot into multiple classes and files
The divisions are more proof of concept than anything, feel free to
change them arround. Also because ~python~ shit might still be broken
that I haven't tested for. Keep an eye out and fix it or yell at me to
fix it.

Also got rid of the last of the VMBotError usage because it was easier
Also fixed an exception when calling faq with no arguments