Fixed bot responding to itself for realsies this time
The base jabberbot.py framework really doesn't work that well for MUCs
and while MUCJabberBot does try to extend it, it's not that great
either. The base framework checks that it's jid is not the incomming
messages jid before responding, but in MUCs the jid takes the form
mucname@server/nickname, whereas the bot thinks its own jid is
username@server.

In the ideal case the bot would keep track of what nicknames it has in
each muc, but it does not. If it did we could comare against that to
reject our own messages, but it doesn't so there's the sloppy compare
against the config file nickname. Someday in the far distant future this
will bite us in the ass but I'm content to wait until then to fix it