Removed the only_direct option from MUCJabberBot
We never used this option (only_direct=False in the bot instantiation) and
even if we did it almost certainly does not do what we want.

If we were to use it, we would want it to respond to its nickname and then a
colon (e.g. Morgooglie: <command>) because that's what tab complete does in
Pidgin. What it actually did was check for the uname or jid without resource
and then a non-aphanumeric. In Morgooglie's case it would have to respond to
jack_haydn@goonfleet.com<space> <command> because the bot runs on Jack's
account.

Finally, the JabberBot callback_presence() returns nothing, so we
shouldn't try to return the return value from it in our own
callback_presence()